Use props instead of duplicate demo html
