--- frontend/src/editor/demoeditor.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<meta
-			name="viewport"
-			content="width=device-width, initial-scale=1.0"
-		/>
-		<title>Editor</title>
-		<style>
-        body, html {
-					font-family: 'Segoe UI', sans-serif, Arial, sans-serif;
-					margin: 0;
-					padding: 0;
-					height: 100%;
-					overflow: hidden;
-        }
-        #container {
-					height: 100%;
-        }
-
-		</style>
-	</head>
-
-	<body>
-		<div id="container"></div>
-	</body>
-</html>
