+++ frontend/src/static/index.html
@@ -56,7 +56,7 @@
 			<h1 id="try-text">Try it out!</h1>
 			<iframe
 				id="editor-frame"
-				src="demoeditor.html"
+				src="editor.html?isDemo=true&wordLimit=500"
 				title="Thoughtful Editor"
 				allowfullscreen
 				allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
